requirements.txt
# Agri-Nexus V1 Platform Dependencies

# Web Framework
streamlit==1.31.0

# AWS SDK
boto3==1.34.34
botocore==1.34.34

# Image Processing
Pillow==10.2.0
opencv-python==4.9.0.80

# Testing
pytest==7.4.4
pytest-asyncio==0.23.4
hypothesis==6.98.3

# Utilities
python-dotenv==1.0.1
requests==2.31.0

# Data Validation
pydantic==2.6.1

# Audio Processing
soundfile==0.12.1
numpy==1.26.3

# Date/Time
python-dateutil==2.8.2
